Moved all contents of .cpp files to .hpp files
For some reason the code that I've been working on started to malfunction. So I decided to follow the books coding syntax
