+++ CMakeLists.txt
@@ -7,5 +7,5 @@ project(FirstRayTracer)
 set(CMAKE_CXX_STANDARD 17)
 set( CMAKE_C_COMPILER  "gcc" )
 
-INCLUDE_DIRECTORIES( include FirstRayTracer/Include )
-add_executable(FirstRayTracer main.cpp SRC/vector3/vector3.cpp SRC/color.cpp SRC/ray/ray.cpp SRC/geometry/sphere.cpp SRC/hittable/hittable_list.cpp)
+INCLUDE_DIRECTORIES( ./Include )
+add_executable(FirstRayTracer main.cpp)
