--- Include/color.h
@@ -1,15 +1,0 @@
-//
-// Created by 유승우 on 2020/05/14.
-//
-
-#ifndef FIRSTRAYTRACER_COLOR_H
-#define FIRSTRAYTRACER_COLOR_H
-
-#include "vector3/vector3.h"
-
-#include <iostream>
-
-void write_color(std::ostream &out, color pixel_color, int samples_per_pixel);
-
-
-#endif //FIRSTRAYTRACER_COLOR_H
