+++ main.cpp
@@ -1,18 +1,23 @@
 #include <iostream>
 
-#include "../Include/rtweekend.h"
-#include "../Include/camera/camera.h"
-#include "../Include/hittable/hittable_list.h"
-#include "../Include/geometry/sphere.h"
-#include "../Include/color.h"
+#include "camera/camera.hpp"
+#include "rtweekend.hpp"
+#include "hittable/hittable_list.hpp"
+#include "geometry/sphere.hpp"
+#include "color.hpp"
 
-color ray_color(const ray& r, const hittable& world)
+color ray_color(const ray& r, const hittable& world, int depth)
 {
     hit_record rec;
+    // If we've exceeded the ray bounce limit, no more light is gathered.
+    if (depth <= 0)
+            return color(0,0,0);
+
     if (world.hit(r, 0, infinity, rec))
     {
+        point3 target = rec.p + rec.normal + random_in_unit_sphere();
         // rec.normal varies from -1 to 1. Thus it must be normalized.
-        return 0.5 * (rec.normal + color(1, 1, 1));
+        return 0.5 * ray_color(ray(rec.p, target - rec.p), world, depth - 1);
     }
     vector3 unit_direction = unit_vector(r.direction());
     auto t = 0.5*(unit_direction.getY() + 1.0);
@@ -29,6 +34,7 @@ int main()
     const int image_width = 384;
     const int image_height = static_cast<int>(image_width / aspect_ratio);
     const int samples_per_pixel = 100;
+    const int max_depth = 50;
 
     std::cout << "P3\n" << image_width << " " << image_height << "\n255\n";
 
@@ -48,7 +54,7 @@ int main()
                 auto u = (i + random_double()) / (image_width - 1);
                 auto v = (j + random_double()) / (image_height - 1);
                 ray r = cam.get_ray(u, v);
-                pixel_color += ray_color(r, world);
+                pixel_color += ray_color(r, world, max_depth);
             }
             write_color(std::cout, pixel_color, samples_per_pixel);
         }
